+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the tool_vault plugin will be documented in this file.
 ### Fixed
 - Removed excessive validation checks, some plugins have `datetime` columns, it is allowed in xmldb
   but reported as error in the "Database check" in Moodle. Vault should allow them.
+- Fixed exceptions when table names contain characters that Moodle does not allow (i.e. `mdl_tablename-old`)
 
 ## [1.5] - 2024-04-07