+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the tool_vault plugin will be documented in this file.
 
 ## Unreleased
 
+### Added
+- Improve how extra/invalid tables are reported, allow to exclude them in one click
 ### Fixed
 - Removed excessive validation checks, some plugins have `datetime` columns, it is allowed in xmldb
   but reported as error in the "Database check" in Moodle. Vault should allow them.